refactoring layout
