+++ sysfunc.h
@@ -12,64 +12,7 @@
    /usr/include.  Until then... */
 
 #if defined(_WIN32) || defined(_WIN64)
-#ifndef _UNISTD_H
-#define _UNISTD_H    1
-
-
-#define index(a,b) strchr(a,b)
-
-/* This is intended as a drop-in replacement for unistd.h on Windows.
- * Please add functionality as neeeded.
- * https://stackoverflow.com/a/826027/1202830
- */
-
-#include <stdlib.h>
-#include <io.h>
-#include <process.h> /* for getpid() and the exec..() family */
-#include <direct.h> /* for _getcwd() and _chdir() */
-
-#define srandom srand
-#define random rand
-
-/* Values for the second argument to access.
-   These may be OR'd together.  */
-#define R_OK    4       /* Test for read permission.  */
-#define W_OK    2       /* Test for write permission.  */
-//#define   X_OK    1       /* execute permission - unsupported in windows*/
-#define F_OK    0       /* Test for existence.  */
-
-#define access _access
-#define dup2 _dup2
-#define execve _execve
-#define ftruncate _chsize
-#define unlink _unlink
-#define fileno _fileno
-#define getcwd _getcwd
-#define chdir _chdir
-#define isatty _isatty
-#define lseek _lseek
-/* read, write, and close are NOT being #defined here, because while there are file handle specific versions for Windows, they probably don't work for sockets. You need to look at your app and consider whether to call e.g. closesocket(). */
-
-#ifdef _WIN64
-#define ssize_t __int64
-#else
-#define ssize_t long
-#endif
-
-#define STDIN_FILENO 0
-#define STDOUT_FILENO 1
-#define STDERR_FILENO 2
-/* should be in some equivalent to <sys/types.h> */
-typedef __int8            int8_t;
-typedef __int16           int16_t; 
-typedef __int32           int32_t;
-typedef __int64           int64_t;
-typedef unsigned __int8   uint8_t;
-typedef unsigned __int16  uint16_t;
-typedef unsigned __int32  uint32_t;
-typedef unsigned __int64  uint64_t;
-
-#endif /* unistd.h  */
+#include "windows/unistd.h"
 #else
 #include <unistd.h>
 #endif
@@ -83,8 +26,8 @@ typedef unsigned __int64  uint64_t;
 #include <memory.h>
 #include <sys/stat.h>
 #if defined(_WIN32) || defined(_WIN64)
-#include "mman.h"
-#include "ya_getopt.h"
+#include "windows/mman.h"
+#include "windows/ya_getopt.h"
 #else
 #include <sys/mman.h>
 #include <getopt.h>
