+++ windows/ya_getopt.c
